[luci] Relocate service CloneNode visit to own file
This will relocate luci service CloneNode visit methods to own file.
